Move ShotoverProcess into test-helpers
