+++ test-helpers/build.rs
@@ -1,0 +1,6 @@
+use std::env;
+
+fn main() {
+    let profile = env::var("PROFILE").unwrap();
+    println!("cargo:rustc-env=PROFILE={profile}");
+}
